Temperature differential trigger for iHost
Scenes on the iHost have limited conditions, and if you want to include a formula you are out of luck. But by running Node-Red on the iHost you can implement more complex logic in Node-Red and do the simple task on the iHost itself.

In this example I show how to create a trigger for a temperature difference. Read two temperature sensors and trigger if one temperature is higher than the other. For example to ventillate/cool a green house and only turn on the fan if the outside temperature is lower than the inside temperature.

https://youtu.be/TXFxgS0h3wk