Consume OWM weather data in Node-Red
I have been using Dark Sky API in the past, but since it got shut down, I migrated over to OWM, which perfectly suits my needs. In this video I show you how to set up to this service and how to query the data in Node-Red. And of course what you get from the service. 

https://youtu.be/e5Rv8V7fTuw