Initial test code
This flow is a model of an irrigation computer which calculates water "intake" from rain and water usage by evaporation based on external temperature. And calculates on a daily basis if watering the garden is needed. More on that in this video: https://www.youtube.com/watch?v=6VkRM6Bu3l0